+++ gaios/AppDelegate.swift
@@ -115,6 +115,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         #endif
 
         AnalyticsManager.shared.countlyStart()
+        applicationWillEnterForeground(application)
 
         return true
     }
@@ -139,22 +140,15 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later.
         // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
         DispatchQueue.global().async {
-            SessionsManager.shared.forEach { (_, session) in
-                if session.connected {
-                    try? session.reconnectHint(hint: ["tor_hint": "disconnect", "hint": "disconnect"])
-                }
-            }
+            SessionsManager.pause()
         }
     }
 
     func applicationWillEnterForeground(_ application: UIApplication) {
         // Called as part of the transition from the background to the active state; here you can undo many of the changes made on entering the background.
         DispatchQueue.global().async {
-            SessionsManager.shared.forEach { (_, session) in
-                if session.connected {
-                    try? session.reconnectHint(hint: ["tor_hint": "connect", "hint": "connect"])
-                }
-            }
+            SessionsManager.resume()
+            AnalyticsManager.shared.setupSession()
         }
     }