+++ gaios/OnBoard/Controllers/WalletSettingsViewController.swift
@@ -1,4 +1,5 @@
 import UIKit
+import PromiseKit
 
 protocol WalletSettingsViewControllerDelegate: AnyObject {
     func didSet(tor: Bool)
@@ -285,7 +286,7 @@ class WalletSettingsViewController: KeyboardViewController {
                 //no change
             }
         }
-
+        updateTorSession(start: switchTor.isOn)
         delegate?.didSet(tor: switchTor.isOn)
         delegate?.didSet(testnet: switchTestnet.isOn)
         dismiss(animated: true, completion: nil)
@@ -298,6 +299,21 @@ class WalletSettingsViewController: KeyboardViewController {
             vc.modalPresentationStyle = .overFullScreen
             vc.disableControls = true
             self.present(vc, animated: true, completion: nil)
+        }
+    }
+
+    func updateTorSession(start: Bool = true) {
+        let bgq = DispatchQueue.global(qos: .background)
+        Guarantee().compactMap(on: bgq) { _ in
+            if start {
+                return TorSessionManager.shared.resume()
+            } else {
+                return TorSessionManager.shared.disconnect()
+            }
+        }.done {
+            AnalyticsManager.shared.setupSession()
+        }.catch { err in
+            print(err.localizedDescription)
         }
     }
 }