+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Enable SPV on multisig
 
 ### Changed
+- Use a global tor session
 
 ### Fixed
 - Update optional fields on jade signing process